+++ sakura_core/charset/CJis.cpp
@@ -448,7 +448,7 @@ EConvertResult CJis::UnicodeToJIS(const CNativeW& cSrc, CMemory* pDstMem)
 
 	// ソースを取得
 	const wchar_t* pSrc = cSrc.GetStringPtr();
-	int nSrcLen = cSrc.GetStringLength();
+	ssize_t nSrcLen = cSrc.GetStringLength();
 
 	// 必要なバッファ容量を確認してバッファを確保
 	char* pDst = new (std::nothrow) char[nSrcLen * 8];
