+++ sakura_core/charset/CUtf7.h
@@ -32,6 +32,6 @@ protected:
 
 	static int _UniToUtf7SetD_block( const wchar_t* pSrc, const ssize_t nSrcLen, char* pDst );
 	static int _UniToUtf7SetB_block( const wchar_t* pSrc, const ssize_t nSrcLen, char* pDst );
-	static int UniToUtf7( const wchar_t* pSrc, const ssize_t nSrcLen, char* pDst, int nDstLen );
+	static ssize_t UniToUtf7( const wchar_t* pSrc, const ssize_t nSrcLen, char* pDst, ssize_t nDstLen );
 };
 #endif /* SAKURA_CUTF7_55498766_1C8A_416B_9F39_88D3D83B8B65_H_ */
