+++ sakura_core/io/CBinaryStream.cpp
@@ -13,19 +13,19 @@ CBinaryInputStream::CBinaryInputStream(LPCWSTR pszFilePath)
 }
 
 //! ストリームの「残り」サイズを取得
-LONGLONG CBinaryInputStream::GetLength()
+ssize_t CBinaryInputStream::GetLength()
 {
 	__int64 nCur = _ftelli64(GetFp());
 	_fseeki64(GetFp(), 0, SEEK_END);
 	__int64 nDataLen = _ftelli64(GetFp());
 	_fseeki64(GetFp(), nCur, SEEK_SET);
-	return (LONGLONG)nDataLen;
+	return static_cast<ssize_t>(nDataLen);
 }
 
 //! データを無変換で読み込む。戻り値は読み込んだバイト数。
-LONGLONG CBinaryInputStream::Read(void* pBuffer, size_t nSizeInBytes)
+ssize_t CBinaryInputStream::Read(void* pBuffer, size_t nSizeInBytes)
 {
-	return (LONGLONG)fread(pBuffer,1,nSizeInBytes,GetFp());
+	return static_cast<ssize_t>(fread(pBuffer,1,nSizeInBytes,GetFp()));
 }
 
 CBinaryOutputStream::CBinaryOutputStream(LPCWSTR pszFilePath, bool bExceptionMode)
